refactor as a package
