+++ testransforms.py
@@ -1,12 +1,10 @@
 import argparse
-import os
 
 from matplotlib import pyplot as plt
-import numpy as np
 from PIL import Image
 
-from videotransforms.video_transforms import Compose, Resize, RandomCrop, RandomRotation, ColorJitter
-from videotransforms.volume_transforms import ClipToTensor
+from torchvideotransforms.video_transforms import Compose, Resize, RandomCrop, RandomRotation, ColorJitter
+from torchvideotransforms.volume_transforms import ClipToTensor
 
 img_path = 'data/cat/cat1.jpeg'
